make a stream helper module
