+++ src/util/mod.rs
@@ -31,6 +31,7 @@ pub mod io_graph;
 pub mod legacy_io;
 pub mod runtime;
 pub mod stdiomux;
+pub mod stream;
 
 /// Create the directory to shove invocation-specific data into, like log files
 /// and sockets.
@@ -76,3 +77,4 @@ pub unsafe fn alloc_uninit_bytes_with_layout(layout: Layout) -> BytesMut {
     };
     BytesMut::from(Bytes::from_owner(mem))
 }
+
